requirements.txt
altgraph==0.17.2
appdirs==1.4.4
apturl==0.5.2
asgiref==3.4.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
ConfigArgParse==1.4.1
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.7.0
django-createsuperuserwithpassword==2.0.0
django-environ==0.4.5
django-heroku==0.3.1
django-rest-auth==0.9.5
django-richtextfield==1.6.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
duplicity==0.8.12.0
entrypoints==0.3
environ==1.0
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-BasicAuth==0.2.0
future==0.18.2
gevent==21.1.2
geventhttpclient==1.4.4
greenlet==1.1.0
gunicorn==20.1.0
httplib2==0.14.0
idna==3.2
importlib-metadata==1.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
locust==1.5.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.2
netifaces==0.10.4
numpy==1.21.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.3.0.38
paramiko==2.6.0
pbr==5.6.0
pdf2image==1.16.0
pdfminer.six==20200517
pdfplumber==0.5.28
pexpect==4.6.0
Pillow==7.0.0
pipenv==2021.5.29
protobuf==3.6.1
psutil==5.8.0
psycopg2==2.9.1
pycairo==1.16.2
pycryptodome==3.10.1
pycups==1.9.73
PyGObject==3.36.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.3.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
pyRFC3339==1.1
PySide2==5.15.2
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.4
pytz==2021.1
pyxdg==0.26
PyYAML==5.3.1
pyzbar==0.1.8
pyzmq==22.1.0
reportlab==3.5.34
requests==2.26.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
shiboken2==5.15.2
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sortedcontainers==2.4.0
sqlparse==0.4.1
systemd-python==234
testresources==2.0.1
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.6
usb-creator==0.3.7
virtualenv==20.4.7
virtualenv-clone==0.5.4
wadllib==1.3.3
Wand==0.6.7
Werkzeug==2.0.1
whitenoise==5.3.0
xkit==0.0.0
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
